+++ pallets/gear/src/ext.rs
@@ -228,7 +228,7 @@ impl EnvExt for LazyPagesExt {
         self.inner.gas(val)
     }
 
-    fn reserve_gas(&mut self, amount: u32, blocks: u32) -> Result<ReservationId, Self::Error> {
+    fn reserve_gas(&mut self, amount: u64, blocks: u32) -> Result<ReservationId, Self::Error> {
         self.inner.reserve_gas(amount, blocks)
     }
